구독하기 구현 완료!!!!!!

+++ src/main/java/com/instagram/dto/SubscribeDto.java
@@ -1,5 +1,7 @@
 package com.instagram.dto;
 
+import java.math.BigInteger;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -13,7 +15,7 @@ public class SubscribeDto {
 
 	private int id;
 	private String username;
-	private String profileImageUrl;
-	private Integer subscribeState;
-	private Integer equalUserState;
+	private String profileImageUrl;	
+	private BigInteger subscribeState;
+	private BigInteger equalUserState;
 }
